Fixed app getting stuck after connecting to bridge
ConnectBulbActivity is no longer called recursively.
Changed ConnectBulbActivity to open the PHHomeActivity for result. This way we can see when the activty is finished and can continue building the project.

The Phillips Hue SDK object where we will control the bridge resides in ProjectSettings and is initialized when launch activity is created. The SDK connects to the bridge earlier. We do not need to use the same Phillips Hue SDK object. I have tested that the bridge information does reach the Phillips Hue SDK object when it is initialized in the projectsettings object in launchactivity.

I commented out code in the HomeActivity so that we don't use the bridge if we have already connected. This is only so we can test the initial set up consecutively.

Changed the listener in PHPushlinkActivty.java so that it finishes the pushlink activity once the bridge is connected.